requirements.txt
flask==3.0.2
numpy==1.26.4
opencv-python>=4.5.0
networkx>=2.6.0
matplotlib>=3.4.0
Pillow==10.2.0
cairosvg>=2.5.0
tqdm==4.66.2
scikit-image==0.21.0
flask-cors==4.0.0 